use Gauge instead of ProgressView
